labeling_tool/server/app/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-mysql==4.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
mysqlclient==1.4.0
PyJWT==2.2.0
pytz==2021.3
sqlparse==0.4.2
